check for nan/inf
with clang 19 / eigen 3.4.0 using isfinite with -ffast-math
results in an error. Disable this particular part of fast-math
to ensure we *can* successfully check for nan where we do.

This flag can have a significant impact on the module performance
and it might be preferred to handle the impact of handed nans per method in the future.